# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SaaS backend with Keycloak-issued JWT auth.\nAll protected endpoints require a Bearer token obtained from Keycloak.",
        "title": "Lightweight SaaS Backend API",
        "contact": {},
        "version": "0.4.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/workspaces/{workspace_id}/connections": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the workspace's connections ordered by name, then id.\nUnlike the workspace listing, the default here is **all**\nstatuses: the operator workflow is draft → verify → activate →\nthe previous one retires, and hiding two thirds of that would\nmake the endpoint unusable. Client secrets are never returned.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connections"
                ],
                "summary": "List a workspace's connections",
                "parameters": [
                    {
                        "type": "string",
                        "example": "ws_5a1b2c3d-4e5f-4a6b-8c9d-0e1f2a3b4c5d",
                        "description": "workspace id",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "draft",
                            "active",
                            "retired",
                            "all"
                        ],
                        "type": "string",
                        "default": "all",
                        "description": "which connections to return",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/connection.ListConnectionsResponse"
                        }
                    },
                    "400": {
                        "description": "invalid_workspace_id, invalid_status_filter",
                        "schema": {
                            "$ref": "#/definitions/connection.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "missing/invalid token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "token lacks admin role",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "workspace_not_found",
                        "schema": {
                            "$ref": "#/definitions/connection.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_error",
                        "schema": {
                            "$ref": "#/definitions/connection.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a connection in **draft** state. The client secret is\nsealed with AES-256-GCM before it reaches the database and is\nnever returned by any endpoint. A draft does nothing until it\nhas been verified and activated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connections"
                ],
                "summary": "Create a connection",
                "parameters": [
                    {
                        "type": "string",
                        "example": "ws_5a1b2c3d-4e5f-4a6b-8c9d-0e1f2a3b4c5d",
                        "description": "workspace id",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "connection to create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/connection.CreateConnectionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/connection.ConnectionResponse"
                        }
                    },
                    "400": {
                        "description": "connection_name_required, connection_base_url_invalid, connection_realm_required, connection_client_id_required, connection_client_secret_required, connection_provider_unsupported, invalid_workspace_id, invalid_request",
                        "schema": {
                            "$ref": "#/definitions/connection.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "missing/invalid token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "token lacks admin role",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "workspace_not_found",
                        "schema": {
                            "$ref": "#/definitions/connection.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "workspace_archived",
                        "schema": {
                            "$ref": "#/definitions/connection.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_error",
                        "schema": {
                            "$ref": "#/definitions/connection.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/workspaces/{workspace_id}/connections/{connection_id}/activate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Promotes a verified draft to **active** and retires the\nworkspace's previous active connection in the same transaction.\nAt most one connection per workspace is active, enforced by a\npartial unique index rather than by application code.\n\nRequires a verification that passed within the last hour.\nUnlike archive, this is deliberately **not** idempotent: a\nrepeat call returns `connection_already_active`, because a\ncaller retrying may believe they are switching away from a\ndifferent connection.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connections"
                ],
                "summary": "Activate a connection",
                "parameters": [
                    {
                        "type": "string",
                        "example": "ws_5a1b2c3d-4e5f-4a6b-8c9d-0e1f2a3b4c5d",
                        "description": "workspace id",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "conn_3f2504e0-4f89-41d3-9a0c-0305e82c3301",
                        "description": "connection id",
                        "name": "connection_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/connection.ConnectionResponse"
                        }
                    },
                    "400": {
                        "description": "invalid_workspace_id, invalid_connection_id",
                        "schema": {
                            "$ref": "#/definitions/connection.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "missing/invalid token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "token lacks admin role",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "workspace_not_found, connection_not_found",
                        "schema": {
                            "$ref": "#/definitions/connection.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "connection_not_verified, connection_verification_expired, connection_already_active, connection_retired, workspace_archived, workspace_has_active_connection",
                        "schema": {
                            "$ref": "#/definitions/connection.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal_error",
                        "schema": {
                            "$ref": "#/definitions/connection.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "connection.CreateConnectionRequest": {
            "type": "object",
            "required": [
                "base_url",
                "client_id",
                "client_secret",
                "name",
                "realm"
            ],
            "properties": {
                "base_url": {
                    "type": "string",
                    "example": "https://kc.example.com"
                },
                "client_id": {
                    "type": "string",
                    "example": "saas-backend-admin"
                },
                "client_secret": {
                    "description": "ClientSecret is write-only. It is sealed with AES-256-GCM before it\nreaches the database and is never returned by any endpoint.",
                    "type": "string",
                    "example": "the-service-account-secret"
                },
                "name": {
                    "type": "string",
                    "example": "Production Keycloak"
                },
                "provider": {
                    "description": "Provider is optional while only one exists; omitted means keycloak.",
                    "type": "string",
                    "example": "keycloak"
                },
                "realm": {
                    "type": "string",
                    "example": "saas"
                }
            }
        },
        "connection.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "connection_not_verified"
                },
                "message": {
                    "type": "string",
                    "example": "Connection must pass verification before it can be activated"
                },
                "request_id": {
                    "type": "string",
                    "example": "3f2504e0-4f89-41d3-9a0c-0305e82c3301"
                }
            }
        },
        "connection.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/connection.ErrorBody"
                }
            }
        },
        "connection.ListConnectionsResponse": {
            "type": "object",
            "properties": {
                "connections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/connection.ConnectionResponse"
                    }
                },
                "count": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "OPERATOR authentication. Type \"Bearer\" followed by a Keycloak-issued access token. The token must carry the realm `admin` role, which is re-checked against the provider on every request.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ProjectKeyAuth": {
            "description": "PROJECT authentication (machine-to-machine). Type \"Bearer\" followed by a project credential, e.g. `Bearer lw_sk_\u003clookup\u003e_\u003csecret\u003e`. The credential is bound to exactly one workspace and carries an explicit scope set; the scope each operation requires is stated in its description. Control-plane operations (workspaces, connections, projects) are operator-only and answer `operator_only` to any credential. Swagger 2.0 cannot express per-operation scopes for an apiKey scheme, so they are documented in prose rather than declared — see docs/PROJECTS.md.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}